requirements.txt
streamlit
pandas
spacy
networkx
pyvis
biopython
google-genai
python-dotenv
